finished client projects task
