+++ src/app/features/add-project/add-project.component.html
@@ -88,7 +88,6 @@
     </div>
     
 
-
     <!-- Submit Button -->
     <button type="submit" [disabled]="projectForm.invalid"
       class="mt-4 bg-blue-600 text-white px-4 py-2 rounded-md hover:bg-blue-700 disabled:bg-gray-400 disabled:cursor-not-allowed">
